requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
sentence-transformers
PyPDF2
python-docx
requests
